feat: Modify rowclass assignment logic in repo listing
In the index.md file, changed the way row_class variable is calculated when listing the repositories. Earlier, it used the cycle function for alternating 'odd' and 'even' class names. However, the approach has been updated to use a modulo operation based on a counter. With this change, the classname toggles between 'odd' and 'even', bringing robustness to the procedure of dynamic class assignment. 🔄👍